move mi tests from test_indexing.py
