+++ pandas/tests/indexing/multiindex/test_setitem.py
@@ -396,3 +396,19 @@ class TestMultiIndexSetItem(object):
         df.loc[ix, "C"] = '_'
 
         assert (df.xs((1, 1))['C'] == '_').all()
+
+    def test_astype_assignment_with_dups(self):
+
+        # GH 4686
+        # assignment with dups that has a dtype change
+        cols = MultiIndex.from_tuples([('A', '1'), ('B', '1'), ('A', '2')])
+        df = DataFrame(np.arange(3).reshape((1, 3)),
+                       columns=cols, dtype=object)
+        index = df.index.copy()
+
+        df['A'] = df['A'].astype(np.float64)
+        tm.assert_index_equal(df.index, index)
+
+        # TODO(wesm): unused variables
+        # result = df.get_dtype_counts().sort_index()
+        # expected = Series({'float64': 2, 'object': 1}).sort_index()
